% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balinski_young.R
\name{app_balinski_young}
\alias{app_balinski_young}
\title{Apportion by the Balinski Young Method}
\usage{
app_balinski_young(size, pop)
}
\arguments{
\item{size}{number of seats to apportion across units}

\item{pop}{a vector of population sizes for each unit}
}
\value{
integer vector
}
\description{
Apportion by the Balinski Young Method
}
\examples{
app_balinski_young(size = 435, pop = state_2020$pop)
}
