% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apportion-package.R
\docType{package}
\name{apportion-package}
\alias{apportion}
\alias{apportion-package}
\title{apportion: Apportion Seats}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Convert populations into integer number of seats for legislative bodies. Implements apportionment methods used historically and currently in the United States for reapportionment after the Census, as described in \url{https://www.census.gov/history/www/reference/apportionment/methods_of_apportionment.html}. These methods are also sometimes used for allocating seats by votes in proportional representation systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christopherkenny/apportion}
  \item \url{http://christophertkenny.com/apportion/}
  \item Report bugs at \url{https://github.com/christopherkenny/apportion/issues}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

Authors:
\itemize{
  \item Cory McCartan \email{mccartan@psu.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID})
}

}
\keyword{internal}
