% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{state_2020}
\alias{state_2020}
\title{2020 State Population Data}
\description{
tibble with columns:
\itemize{
\item \code{GEOID}: Federal Information Processing Standards codes
\item \code{name}: title case state name
\item \code{pop}: 2020 population
\item \code{abb}: two letter postal abbreviations
}
}
\examples{
head(state_2020)
}
\concept{data}
