% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adams.R
\name{app_adams}
\alias{app_adams}
\title{Apportion by the Adams Method}
\usage{
app_adams(size, pop)
}
\arguments{
\item{size}{An integer number of seats to apportion across units}

\item{pop}{A vector of population sizes for each unit}
}
\value{
An integer vector of the same length as \code{pop} with the number of seats apportioned to each unit.
}
\description{
Apportion by the Adams Method
}
\examples{
app_adams(size = 435, pop = state_2020$pop)
}
