% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adams.R
\name{app_adams}
\alias{app_adams}
\title{Apportion by the Adams Method}
\usage{
app_adams(size, pop)
}
\arguments{
\item{size}{number of seats to apportion across units}

\item{pop}{a vector of population sizes for each unit}
}
\value{
integer vector
}
\description{
Apportion by the Adams Method
}
\examples{
app_adams(size = 435, pop = state_2020$pop)
}
