% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dean.R
\name{app_dean}
\alias{app_dean}
\title{Apportion by the Dean Method}
\usage{
app_dean(size, pop)
}
\arguments{
\item{size}{number of seats to apportion across units}

\item{pop}{a vector of population sizes for each unit}
}
\value{
integer vector
}
\description{
Apportion by the Dean Method
}
\examples{
app_dean(size = 435, pop = state_2020$pop)
}
