% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhondt.R
\name{app_dhondt}
\alias{app_dhondt}
\title{Apportion by the D'Hondt Method}
\usage{
app_dhondt(size, pop)
}
\arguments{
\item{size}{number of seats to apportion across units}

\item{pop}{a vector of population sizes for each unit}
}
\value{
integer vector
}
\description{
Apportion by the D'Hondt Method
}
\examples{
app_dhondt(size = 435, pop = state_2020$pop)
}
