% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamilton_vinton.R
\name{app_hamilton_vinton}
\alias{app_hamilton_vinton}
\title{Apportion by the Hamilton-Vinton Method}
\usage{
app_hamilton_vinton(size, pop)
}
\arguments{
\item{size}{number of seats to apportion across units}

\item{pop}{a vector of population sizes for each unit}
}
\value{
integer vector
}
\description{
Apportion by the Hamilton-Vinton Method
}
\examples{
app_hamilton_vinton(size = 435, pop = state_2020$pop)
}
