% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huntington_hill.R
\name{app_huntington_hill}
\alias{app_huntington_hill}
\title{Apportion by the Huntington-Hill Method}
\usage{
app_huntington_hill(size, pop)
}
\arguments{
\item{size}{number of seats to apportion across units}

\item{pop}{a vector of population sizes for each unit}
}
\value{
integer vector
}
\description{
Apportion by the Huntington-Hill Method
}
\examples{
app_huntington_hill(size = 435, pop = state_2020$pop)
}
