% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jefferson.R
\name{app_jefferson}
\alias{app_jefferson}
\title{Apportion by the Jefferson Method}
\usage{
app_jefferson(size, pop)
}
\arguments{
\item{size}{number of seats to apportion across units}

\item{pop}{a vector of population sizes for each unit}
}
\value{
integer vector
}
\description{
Apportion by the Jefferson Method
}
\examples{
app_jefferson(size = 435, pop = state_2020$pop)
}
