% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webster.R
\name{app_webster}
\alias{app_webster}
\title{Apportion by the Webster Method}
\usage{
app_webster(size, pop)
}
\arguments{
\item{size}{An integer number of seats to apportion across units}

\item{pop}{A vector of population sizes for each unit}
}
\value{
integer vector
}
\description{
Apportion by the Webster Method
}
\examples{
app_webster(size = 435, pop = state_2020$pop)
}
