% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{state_2020}
\alias{state_2020}
\title{2020 State Population Data}
\description{
tibble with columns:
\itemize{
\item \code{GEOID}: Federal Information Processing Standards code
\item \code{name}: State name
\item \code{pop}: 2020 census population
\item \code{abb}: Two-letter postal abbreviation
}
}
\examples{
head(state_2020)
}
\concept{data}
