% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{state_2020}
\alias{state_2020}
\title{state_2020 (2020 State Data)}
\usage{
data('state_2020')
}
\value{
tibble with state identifying information
}
\description{
tibble with columns:
\itemize{
\item GEOID: Federal Information Processing Standards codes
\item name: title case state name
\item pop: 2020 population
\item abb: two letter postal abbreviations
}
}
\examples{
data('state_2020')
}
\concept{data}
